# Improve performance of json lookups in case of errors