--- a/benchmarks/src/main/scala/play/api/libs/json/JsObjectBench.scala
+++ b/benchmarks/src/main/scala/play/api/libs/json/JsObjectBench.scala
@@ -1,3 +1,7 @@
+/*
+ * Copyright (C) from 2022 The Play Framework Contributors <https://github.com/playframework>, 2011-2021 Lightbend Inc. <https://www.lightbend.com>
+ */
+
 package play.api.libs.json
 
 import org.openjdk.jmh.annotations._